requirements.txt
openai
python-dotenv
yfinance
